requirements.txt
duckdb>=1.1.0
feedparser
pandas
requests
pyyaml
python-dateutil
dateparser
spacy>=3.7
geopy
trafilatura
newspaper3k
simhash
streamlit
# After install run: python -m spacy download en_core_web_sm
